requirements.txt
pandas
matplotlib
seaborn
ipywidgets
ipykernel
ipython
pillow
numpy
opencv-python
sklearn
